requirements.txt
blis==0.7.4
boto3==1.17.78
botocore==1.20.78
Brotli==1.0.9
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==2.0.1
Flask-Compress==1.9.0
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathy==0.5.2
patsy==0.5.1
pickle-mixin==1.0.2
Pillow==8.2.0
plotly==4.14.3
plotly-express==0.4.1
preshed==3.0.5
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
sklearn==0.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
statsmodels==0.12.2
thinc==8.0.3
threadpoolctl==2.1.0
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
Werkzeug==2.0.1
wordcloud==1.8.1
